requirements.txt
python-multipart
fastapi==0.111.0
pydantic-settings==2.2.1
pydantic==2.7.1
streamlit
instructor
azure-search-documents
azure-storage-blob
